Split StationSheet into focused station components
StationSheet.tsx (850 lines at its peak) is now a 98-line shell + body
router; the pieces live beside it:

- useOperationProgress: the one derivation of phases/current/waiting/
  progress, shared by the status line and both sheet bodies — the three
  used to compute phase math independently with subtle differences
  (DirectFeed's progress ignored later phases; identical today since
  direct-feed ops are single-phase, but now it can't drift)
- ParameterScaleRow: the labeled settings scale (Z chip, cut-line vs
  detent presentation), shared by both bodies; the value fallback now
  consistently honors param.defaultValue (benches declare none today)
- SlotDiagram: the bench's input->output bay with its four output
  fallback tiers
- VerbRow/OutputsRow, StatusText, DirectFeedSheet, BenchSheet,
  ContentsSheet: extracted verbatim
